Move proto build into crate
