+++ build.rs
@@ -1,27 +1,15 @@
 // #![allow(clippy::all)]
 use ethers::contract::Abigen;
-use std::process;
-
-use std::path::Path;
-use std::{
-    fs::{self, create_dir_all, remove_dir_all},
-    path::PathBuf,
-};
-
 use merkle_hash::{Algorithm, MerkleTree};
-use walkdir::WalkDir;
+use std::process;
+use std::fs;
 
-/// A temporary directory for proto building
-const TMP_PATH: &str = "/tmp/steward/";
-/// the output directory
-const OUT_PATH: &str = "crates/steward-proto/src/gen/";
+
 const HASH_ABI_FILE: &str = "hash_abi";
-const HASH_PROTO_FILE: &str = "hash_proto";
 
 fn main() {
     // only generate bindings if changes have occurred
     let previous_abi_hash = fs::read_to_string(HASH_ABI_FILE).unwrap_or_default();
-    let previous_proto_hash = fs::read_to_string(HASH_PROTO_FILE).unwrap_or_default();
     let current_abi_hash = MerkleTree::builder("abi")
         .algorithm(Algorithm::Blake3)
         .hash_names(false)
@@ -30,25 +18,11 @@ fn main() {
         .root
         .item
         .hash;
-    let current_proto_hash = MerkleTree::builder("proto")
-        .algorithm(Algorithm::Blake3)
-        .hash_names(false)
-        .build()
-        .expect("proto dir merkle tree build failed")
-        .root
-        .item
-        .hash;
     let current_abi_hash = hex::encode(current_abi_hash);
-    let current_proto_hash = hex::encode(current_proto_hash);
 
     if current_abi_hash != previous_abi_hash {
         generate_contract_abis();
         fs::write(HASH_ABI_FILE, current_abi_hash).expect("failed to write abi hash");
-    }
-
-    if current_proto_hash != previous_proto_hash {
-        generate_rust_protos();
-        fs::write(HASH_PROTO_FILE, current_proto_hash).expect("failed to write proto hash");
     }
 }
 
@@ -157,86 +131,4 @@ fn generate_contract_abis() {
             Err(e) => println!("Error writing {}.rs: {}", file_name, e),
         }
     })
-}
-
-fn generate_rust_protos() {
-    let out_dir = Path::new(&OUT_PATH);
-    let tmp_dir = Path::new(&TMP_PATH);
-    let root = env!("CARGO_MANIFEST_DIR");
-    let root: PathBuf = root.parse().unwrap();
-    let mut steward_proto_dir = root;
-    steward_proto_dir.push("proto/steward/v4");
-    let steward_proto_dir = [steward_proto_dir];
-
-    // List available proto files
-    let mut protos: Vec<PathBuf> = vec![];
-    for proto_path in &steward_proto_dir {
-        protos.append(
-            &mut WalkDir::new(proto_path)
-                .into_iter()
-                .filter_map(|e| e.ok())
-                .filter(|e| {
-                    e.file_type().is_file()
-                        && e.path().extension().is_some()
-                        && e.path().extension().unwrap() == "proto"
-                })
-                .map(|e| e.into_path())
-                .collect(),
-        );
-    }
-
-    // create directories for temporary build dirs
-    fs::create_dir_all(tmp_dir)
-        .unwrap_or_else(|_| panic!("Failed to create {:?}", tmp_dir.to_str()));
-
-    // Compile all proto files
-    let mut config = prost_build::Config::default();
-    config.out_dir(tmp_dir);
-    config.type_attribute(".", "#[derive(serde::Deserialize, serde::Serialize)]");
-    config.compile_protos(&protos, &steward_proto_dir).unwrap();
-
-    // Compile all proto client for GRPC services
-    println!("Compiling proto clients for GRPC services...");
-    tonic_build::configure()
-        .build_client(true)
-        .build_server(true)
-        .file_descriptor_set_path(tmp_dir.join("descriptor.bin"))
-        .out_dir(tmp_dir)
-        .compile_protos_with_config(config, &protos, &steward_proto_dir)
-        .unwrap();
-
-    copy_generated_files(tmp_dir, out_dir);
-    println!("Done!");
-}
-
-fn copy_generated_files(from_dir: &Path, to_dir: &Path) {
-    println!("Copying generated files into '{}'...", to_dir.display());
-
-    // Remove old compiled files
-    remove_dir_all(to_dir).unwrap_or_default();
-    create_dir_all(to_dir).unwrap();
-
-    let mut filenames = Vec::new();
-
-    // Copy new compiled files (prost does not use folder structures)
-    let errors = WalkDir::new(from_dir)
-        .into_iter()
-        .filter_map(|e| e.ok())
-        .filter(|e| e.file_type().is_file())
-        .map(|e| {
-            let filename = e.file_name().to_os_string().to_str().unwrap().to_string();
-            filenames.push(filename.clone());
-            let contents = fs::read(e.path())?;
-            fs::write(format!("{}/{}", to_dir.display(), &filename), contents)
-        })
-        .filter_map(|e| e.err())
-        .collect::<Vec<_>>();
-
-    if !errors.is_empty() {
-        for e in errors {
-            eprintln!("[error] Error while copying compiled file: {}", e);
-        }
-
-        panic!("[error] Aborted.");
-    }
 }
